Добавить companyController
