+++ src/server/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"encoding/json"
+	"github.com/KondratenkoAndrey/alef/controllers"
 	"github.com/gorilla/mux"
 	"log"
 	"net/http"
@@ -9,29 +9,6 @@ import (
 
 func main() {
 	router := mux.NewRouter()
-	router.HandleFunc("/company-info", CompanyInfoHandler)
+	router.HandleFunc("/company-info", controllers.CompanyInfoHandler)
 	log.Fatal(http.ListenAndServe(":8080", router))
-}
-
-func CompanyInfoHandler(w http.ResponseWriter, r *http.Request) {
-	log.Println(r.Method, r.URL)
-	w.Header().Set("Content-Type", "application/json")
-	pages := getCompanyInfo()
-	json.NewEncoder(w).Encode(pages)
-}
-
-func getCompanyInfo() CompanyInfo {
-	return CompanyInfo {
-		ShortName: "Алеф",
-		FullName: "Компания Алеф",
-		Phone: "+7(789)123-45-67",
-		Email: "info@example.ru",
-	}
-}
-
-type CompanyInfo struct {
-	ShortName	string `json:"shortName"`
-	FullName	string `json:"fullName"`
-	Phone		string `json:"phone"`
-	Email		string `json:"email"`
 }
